Convert posts and delete in jQuery(AJAX)

+++ controllers/posts_controller.js
@@ -3,11 +3,20 @@ const Comment = require('../models/comment');
 
 module.exports.create = async function(req, res){
     try{
-        await Post.create({
+        let post = await Post.create({
             content: req.body.content,
             user: req.user._id
         });
         
+        if (req.xhr){
+            return res.status(200).json({
+                data: {
+                    post: post
+                },
+                message: "Post created!"
+            });
+        }
+
         req.flash('success', 'Post published!');
         return res.redirect('back');
 
@@ -19,27 +28,30 @@ module.exports.create = async function(req, res){
 }
 
 
-module.exports.destroy = async function(req, res){
+module.exports.destroy = async (req, res) => {
+    try {
 
-    try{
-        let post = await Post.findById(req.params.id);
+        const post = await Post.findById(req.params.id);
 
-        if (post.user == req.user.id){
-            post.remove();
+        if (post && post.user == req.user.id) {
+            await Post.findByIdAndDelete(post.id);
+            await Comment.deleteMany({ post: req.params.id })
 
-            await Comment.deleteMany({post: req.params.id});
+            if(req.xhr){
+                return res.status(200).json({
+                    data: {
+                        post_id: req.params.id
+                    },
+                    message: 'post deleted!'
+                })
+            }
 
-            req.flash('success', 'Post and associated comments deleted!');
-
-            return res.redirect('back');
-        }else{
-            req.flash('error', 'You cannot delete this post!');
-            return res.redirect('back');
+            req.flash('success', 'post deleted')
+            res.redirect('back');
         }
 
-    }catch(err){
+    } catch (err) {
         req.flash('error', err);
-        return res.redirect('back');
+        res.redirect('back');
     }
-    
 }
